perf(app): lazy-load GraphExplorerGPU to drop 702KB cosmos chunk from eager startup (TRA-1185)
Code-splits GraphExplorerGPU behind React.lazy and eliminates the cosmos manual
chunk in packages/app/vite.config.ts. Previously, manualChunks forced Rollup to
place Vite's dynamic-import preload helper into the cosmos chunk; because index.js
needed that helper, index.html emitted a modulepreload link for cosmos-*.js (702 KB),
forcing the full WebGL/Cosmos stack to be fetched and parsed on startup.

Measurements (3 samples, darwin 25.5.0 arm64, node 22.22.3):
- renderer_eager_kb: 2084 KB -> 1358 KB (-726 KB, -34.8% vs last baseline; -35.7% vs 2112 KB pre-fix)
- renderer_first_content_ms: 93 ms -> 89 ms (-4.3%)
- renderer_bundle_kb: 2346 KB -> 2379 KB (+1.4%, flat; bytes moved from eager to lazy chunk)
- cold_start_ms: 436 ms
- heap_idle_mb: 9.5 MB

Also:
- Added retry focus logic in openFileInGraph so file navigation waits for lazy GraphExplorerGPU to mount.
- Extracted GraphExplorerGPUHandle and settings types into graph-types.ts to avoid circular lazy imports.
- Added Savings and GraphExplorerGPU assertions to lazy-tabs.test.ts.
- Added Savings tab to perf-marks.test.ts and perf-screens.mjs.
- Appended measured run to docs/perf/baseline.json and docs/perf/screens.json; updated docs/perf/README.md.
